refactor base launch to use simple_launch package
