+++ nightwing/platform_common/config/log_topics.yaml
@@ -1,5 +1,17 @@
 topics:
-  # Vision topics
+  # Perception topics
   - /depth/image_rect_compressed
   - /color/image_compressed
-  # Odometry topics
+  # Topics from PX4
+  - /fmu/out/battery_status
+  - /fmu/out/estimator_status_flags
+  - /fmu/out/manual_control_setpoint
+  - /fmu/out/position_setpoint_triplet
+  - /fmu/out/sensor_combined
+  - /fmu/out/vehicle_attitude
+  - /fmu/out/vehicle_global_position
+  - /fmu/out/vehicle_gps_position
+  - /fmu/out/vehicle/_local_position
+  - /fmu/out/vehicle_odometry
+  - /fmu/out/vehicle_status
+  - /fmu/out/vehicle_trajectory_waypoint_desired
