+++ nightwing/platform_common/config/perception/realsense.yaml
@@ -1,0 +1,38 @@
+device_type: ''
+serial_no: ''
+usb_port_id: ''
+
+rgb_camera:
+  profile: '1280x720x15'
+  auto_exposure_priority: false
+depth_module:
+  profile: '1280x720x15'
+
+# -----------------------
+# Sensor Enabling
+# -----------------------
+# we want rgbd but you need to enable the 
+# other two for it work
+enable_color: true
+enable_depth: true
+enable_rgbd: true
+
+enable_accel: false
+enable_gyro: false
+enable_infra1: false
+enable_infra2: false
+pointcloud:
+  enable: false
+# ----------------------
+
+intra_process_comms: true
+enable_sync: true
+
+# needs to be enabled for rgbd
+align_depth:
+  enable: true
+
+# depth colorizer so we can apply h.264 encoding
+# to it for data logging
+colorizer:
+  enable: true
