--- nightwing/platform_common/config/vision/realsense.yaml
@@ -1,26 +1,0 @@
-device_type: ''
-serial_no: ''
-usb_port_id: ''
-
-rgb_camera:
-  profile: '1280x720x30'
-
-depth_module:
-  profile: '1280x720x30'
-
-enable_accel: false
-enable_color: true
-enable_depth: true
-enable_rgbd: true
-enable_gyro: false
-enable_infra1: false
-enable_infra2: false
-
-intra_process_comms: true
-
-pointcloud:
-  enable: false
-
-enable_sync: true
-align_depth:
-  enable: true
